e-commerce-microservices/services/auth-service/requirements.txt
Django>=4.2.0,<5.0
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.3.0
django-cors-headers>=4.3.0
psycopg2-binary>=2.9.9
PyJWT>=2.8.0
python-dotenv>=1.0.0
drf-yasg>=1.21.7
gunicorn>=21.2.0
